Fix jump-past-header on reach page smoothscroll
Pushing the hash into the URL was making the page jump down to that ID, which was too far with
the fixed header. With this change, the back button does not return the user back to the top of
the reach page, but that may have been surprising behaviour anyway. It now takes the user back
to the page they were on before the reach page.